LOG4J2-1344 added method ByteBufferDestination.size() - count total bytes written to destination
The size() method is necessary to support file rollover (file size trigger):
Managers for Rolling...Appenders currently override the write() method and count the number of written bytes there.
When the Layout encodes the LogEvent directly into the ByteBufferDestination, the write() method is not called.
The ByteBufferDestination.size() gives the Manager enough information to decide if a rollover is required.